+++ test/two-dim.c
@@ -160,7 +160,7 @@ static void check_data_3D ( double ***data, int rank, int *dim,
 			    int hwidth, int nc, int *neighbors ) 
 {
     int i, j, k, lres=1, gres;
-    int should_be;
+    double should_be;
 
     should_be = neighbors[0]==MPI_PROC_NULL ? -1 : neighbors[0];
     for ( j=hwidth; j<dim[1]-hwidth; j++ ) {
@@ -241,7 +241,7 @@ static void check_data_2D ( double **data, int rank, int *dim, int hwidth,
 			    int *neighbors)
 {
     int i, j, lres=1, gres;
-    int should_be;
+    double should_be;
 
 
     should_be = neighbors[0]==MPI_PROC_NULL ? -1 : neighbors[0];
